Crate rusoto_comprehend

Amazon Comprehend is an AWS service for gaining insight into the content of documents. Use these actions to determine the topics contained in your documents, the topics they discuss, the predominant sentiment expressed in them, the predominant language used, and more.

If you're using the service, you're probably looking for ComprehendClient and Comprehend.

Structs

BatchDetectDominantLanguageItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectDominantLanguageRequest

BatchDetectDominantLanguageResponse

BatchDetectEntitiesItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectEntitiesRequest

BatchDetectEntitiesResponse

BatchDetectKeyPhrasesItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectKeyPhrasesRequest

BatchDetectKeyPhrasesResponse

BatchDetectSentimentItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectSentimentRequest

BatchDetectSentimentResponse

BatchDetectSyntaxItemResult
The result of calling the operation. The operation returns one object that is successfully processed by the operation.

BatchDetectSyntaxRequest

BatchDetectSyntaxResponse

BatchItemError
Describes an error that occurred while processing a document in a batch. The operation returns on BatchItemError object for each document that contained an error.

ClassifierEvaluationMetrics
Describes the result metrics for the test data associated with an documentation classifier.

ClassifierMetadata
Provides information about a document classifier.

ComprehendClient
A client for the Amazon Comprehend API.

CreateDocumentClassifierRequest

CreateDocumentClassifierResponse

CreateEntityRecognizerRequest

CreateEntityRecognizerResponse

DeleteDocumentClassifierRequest

DeleteDocumentClassifierResponse

DeleteEntityRecognizerRequest

DeleteEntityRecognizerResponse

DescribeDocumentClassificationJobRequest

DescribeDocumentClassificationJobResponse

DescribeDocumentClassifierRequest

DescribeDocumentClassifierResponse

DescribeDominantLanguageDetectionJobRequest

DescribeDominantLanguageDetectionJobResponse

DescribeEntitiesDetectionJobRequest

DescribeEntitiesDetectionJobResponse

DescribeEntityRecognizerRequest

DescribeEntityRecognizerResponse

DescribeKeyPhrasesDetectionJobRequest

DescribeKeyPhrasesDetectionJobResponse

DescribeSentimentDetectionJobRequest

DescribeSentimentDetectionJobResponse

DescribeTopicsDetectionJobRequest

DescribeTopicsDetectionJobResponse

DetectDominantLanguageRequest

DetectDominantLanguageResponse

DetectEntitiesRequest

DetectEntitiesResponse

DetectKeyPhrasesRequest

DetectKeyPhrasesResponse

DetectSentimentRequest

DetectSentimentResponse

DetectSyntaxRequest

DetectSyntaxResponse

DocumentClassificationJobFilter
Provides information for filtering a list of document classification jobs. For more information, see the operation. You can provide only one filter parameter in each request.

DocumentClassificationJobProperties
Provides information about a document classification job.

DocumentClassifierFilter
Provides information for filtering a list of document classifiers. You can only specify one filtering parameter in a request. For more information, see the operation.

DocumentClassifierInputDataConfig
The input properties for training a document classifier.

For more information on how the input file is formatted, see how-document-classification-training-data.

DocumentClassifierOutputDataConfig
Provides output results configuration parameters for custom classifier jobs.

DocumentClassifierProperties
Provides information about a document classifier.

DominantLanguage
Returns the code for the dominant language in the input text and the level of confidence that Amazon Comprehend has in the accuracy of the detection.

DominantLanguageDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

DominantLanguageDetectionJobProperties
Provides information about a dominant language detection job.

EntitiesDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

EntitiesDetectionJobProperties
Provides information about an entities detection job.

Entity
Provides information about an entity.

EntityRecognizerAnnotations
Describes the annotations associated with a entity recognizer.

EntityRecognizerDocuments
Describes the training documents submitted with an entity recognizer.

EntityRecognizerEntityList
Describes the entity recognizer submitted with an entity recognizer.

EntityRecognizerEvaluationMetrics
Detailed information about the accuracy of an entity recognizer.

EntityRecognizerFilter
Provides information for filtering a list of entity recognizers. You can only specify one filtering parameter in a request. For more information, see the operation./>

EntityRecognizerInputDataConfig
Specifies the format and location of the input data.

EntityRecognizerMetadata
Detailed information about an entity recognizer.

EntityRecognizerMetadataEntityTypesListItem
Individual item from the list of entity types in the metadata of an entity recognizer.

EntityRecognizerProperties
Describes information about an entity recognizer.

EntityTypesListItem
Information about an individual item on a list of entity types.

InputDataConfig
The input properties for a topic detection job.

KeyPhrase
Describes a key noun phrase.

KeyPhrasesDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

KeyPhrasesDetectionJobProperties
Provides information about a key phrases detection job.

ListDocumentClassificationJobsRequest

ListDocumentClassificationJobsResponse

ListDocumentClassifiersRequest

ListDocumentClassifiersResponse

ListDominantLanguageDetectionJobsRequest

ListDominantLanguageDetectionJobsResponse

ListEntitiesDetectionJobsRequest

ListEntitiesDetectionJobsResponse

ListEntityRecognizersRequest

ListEntityRecognizersResponse

ListKeyPhrasesDetectionJobsRequest

ListKeyPhrasesDetectionJobsResponse

ListSentimentDetectionJobsRequest

ListSentimentDetectionJobsResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

ListTopicsDetectionJobsRequest

ListTopicsDetectionJobsResponse

OutputDataConfig

Provides configuration parameters for the output of topic detection jobs.

PartOfSpeechTag
Identifies the part of speech represented by the token and gives the confidence that Amazon Comprehend has that the part of speech was correctly identified. For more information about the parts of speech that Amazon Comprehend can identify, see how-syntax.

SentimentDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

SentimentDetectionJobProperties
Provides information about a sentiment detection job.

SentimentScore
Describes the level of confidence that Amazon Comprehend has in the accuracy of its detection of sentiments.

StartDocumentClassificationJobRequest

StartDocumentClassificationJobResponse

StartDominantLanguageDetectionJobRequest

StartDominantLanguageDetectionJobResponse

StartEntitiesDetectionJobRequest

StartEntitiesDetectionJobResponse

StartKeyPhrasesDetectionJobRequest

StartKeyPhrasesDetectionJobResponse

StartSentimentDetectionJobRequest

StartSentimentDetectionJobResponse

StartTopicsDetectionJobRequest

StartTopicsDetectionJobResponse

StopDominantLanguageDetectionJobRequest

StopDominantLanguageDetectionJobResponse

StopEntitiesDetectionJobRequest

StopEntitiesDetectionJobResponse

StopKeyPhrasesDetectionJobRequest

StopKeyPhrasesDetectionJobResponse

StopSentimentDetectionJobRequest

StopSentimentDetectionJobResponse

StopTrainingDocumentClassifierRequest

StopTrainingDocumentClassifierResponse

StopTrainingEntityRecognizerRequest

StopTrainingEntityRecognizerResponse

SyntaxToken
Represents a work in the input text that was recognized and assigned a part of speech. There is one syntax token record for each word in the source text.

Tag
A key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For example, a tag with the key-value pair ‘Department’:’Sales’ might be added to a resource to indicate its use by a particular department.

TagResourceRequest

TagResourceResponse

TopicsDetectionJobFilter
Provides information for filtering topic detection jobs. For more information, see .

TopicsDetectionJobProperties
Provides information about a topic detection job.

UntagResourceRequest

UntagResourceResponse

VpcConfig
Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for the job. For For more information, see Amazon VPC.

Enums

BatchDetectDominantLanguageError
Errors returned by BatchDetectDominantLanguage

BatchDetectEntitiesError
Errors returned by BatchDetectEntities

BatchDetectKeyPhrasesError
Errors returned by BatchDetectKeyPhrases

BatchDetectSentimentError
Errors returned by BatchDetectSentiment

BatchDetectSyntaxError
Errors returned by BatchDetectSyntax

CreateDocumentClassifierError
Errors returned by CreateDocumentClassifier

CreateEntityRecognizerError
Errors returned by CreateEntityRecognizer

DeleteDocumentClassifierError
Errors returned by DeleteDocumentClassifier

DeleteEntityRecognizerError
Errors returned by DeleteEntityRecognizer

DescribeDocumentClassificationJobError
Errors returned by DescribeDocumentClassificationJob

DescribeDocumentClassifierError
Errors returned by DescribeDocumentClassifier

DescribeDominantLanguageDetectionJobError
Errors returned by DescribeDominantLanguageDetectionJob

DescribeEntitiesDetectionJobError
Errors returned by DescribeEntitiesDetectionJob

DescribeEntityRecognizerError
Errors returned by DescribeEntityRecognizer

DescribeKeyPhrasesDetectionJobError
Errors returned by DescribeKeyPhrasesDetectionJob

DescribeSentimentDetectionJobError
Errors returned by DescribeSentimentDetectionJob

DescribeTopicsDetectionJobError
Errors returned by DescribeTopicsDetectionJob

DetectDominantLanguageError
Errors returned by DetectDominantLanguage

DetectEntitiesError
Errors returned by DetectEntities

DetectKeyPhrasesError
Errors returned by DetectKeyPhrases

DetectSentimentError
Errors returned by DetectSentiment

DetectSyntaxError
Errors returned by DetectSyntax

ListDocumentClassificationJobsError
Errors returned by ListDocumentClassificationJobs

ListDocumentClassifiersError
Errors returned by ListDocumentClassifiers

ListDominantLanguageDetectionJobsError
Errors returned by ListDominantLanguageDetectionJobs

ListEntitiesDetectionJobsError
Errors returned by ListEntitiesDetectionJobs

ListEntityRecognizersError
Errors returned by ListEntityRecognizers

ListKeyPhrasesDetectionJobsError
Errors returned by ListKeyPhrasesDetectionJobs

ListSentimentDetectionJobsError
Errors returned by ListSentimentDetectionJobs

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTopicsDetectionJobsError
Errors returned by ListTopicsDetectionJobs

StartDocumentClassificationJobError
Errors returned by StartDocumentClassificationJob

StartDominantLanguageDetectionJobError
Errors returned by StartDominantLanguageDetectionJob

StartEntitiesDetectionJobError
Errors returned by StartEntitiesDetectionJob

StartKeyPhrasesDetectionJobError
Errors returned by StartKeyPhrasesDetectionJob

StartSentimentDetectionJobError
Errors returned by StartSentimentDetectionJob

StartTopicsDetectionJobError
Errors returned by StartTopicsDetectionJob

StopDominantLanguageDetectionJobError
Errors returned by StopDominantLanguageDetectionJob

StopEntitiesDetectionJobError
Errors returned by StopEntitiesDetectionJob

StopKeyPhrasesDetectionJobError
Errors returned by StopKeyPhrasesDetectionJob

StopSentimentDetectionJobError
Errors returned by StopSentimentDetectionJob

StopTrainingDocumentClassifierError
Errors returned by StopTrainingDocumentClassifier

StopTrainingEntityRecognizerError
Errors returned by StopTrainingEntityRecognizer

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

Traits

Comprehend
Trait representing the capabilities of the Amazon Comprehend API. Amazon Comprehend clients implement this trait.