Crate rusoto_comprehend

Amazon Comprehend is an AWS service for gaining insight into the content of documents. Use these actions to determine the topics contained in your documents, the topics they discuss, the predominant sentiment expressed in them, the predominant language used, and more.

If you're using the service, you're probably looking for ComprehendClient and Comprehend.

Structs

BatchDetectDominantLanguageItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectDominantLanguageRequest

BatchDetectDominantLanguageResponse

BatchDetectEntitiesItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectEntitiesRequest

BatchDetectEntitiesResponse

BatchDetectKeyPhrasesItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectKeyPhrasesRequest

BatchDetectKeyPhrasesResponse

BatchDetectSentimentItemResult
The result of calling the operation. The operation returns one object for each document that is successfully processed by the operation.

BatchDetectSentimentRequest

BatchDetectSentimentResponse

BatchDetectSyntaxItemResult
The result of calling the operation. The operation returns one object that is successfully processed by the operation.

BatchDetectSyntaxRequest

BatchDetectSyntaxResponse

BatchItemError
Describes an error that occurred while processing a document in a batch. The operation returns on BatchItemError object for each document that contained an error.

ComprehendClient
A client for the Amazon Comprehend API.

DescribeDominantLanguageDetectionJobRequest

DescribeDominantLanguageDetectionJobResponse

DescribeEntitiesDetectionJobRequest

DescribeEntitiesDetectionJobResponse

DescribeKeyPhrasesDetectionJobRequest

DescribeKeyPhrasesDetectionJobResponse

DescribeSentimentDetectionJobRequest

DescribeSentimentDetectionJobResponse

DescribeTopicsDetectionJobRequest

DescribeTopicsDetectionJobResponse

DetectDominantLanguageRequest

DetectDominantLanguageResponse

DetectEntitiesRequest

DetectEntitiesResponse

DetectKeyPhrasesRequest

DetectKeyPhrasesResponse

DetectSentimentRequest

DetectSentimentResponse

DetectSyntaxRequest

DetectSyntaxResponse

DominantLanguage
Returns the code for the dominant language in the input text and the level of confidence that Amazon Comprehend has in the accuracy of the detection.

DominantLanguageDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

DominantLanguageDetectionJobProperties
Provides information about a dominant language detection job.

EntitiesDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

EntitiesDetectionJobProperties
Provides information about an entities detection job.

Entity
Provides information about an entity.

InputDataConfig
The input properties for a topic detection job.

KeyPhrase
Describes a key noun phrase.

KeyPhrasesDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

KeyPhrasesDetectionJobProperties
Provides information about a key phrases detection job.

ListDominantLanguageDetectionJobsRequest

ListDominantLanguageDetectionJobsResponse

ListEntitiesDetectionJobsRequest

ListEntitiesDetectionJobsResponse

ListKeyPhrasesDetectionJobsRequest

ListKeyPhrasesDetectionJobsResponse

ListSentimentDetectionJobsRequest

ListSentimentDetectionJobsResponse

ListTopicsDetectionJobsRequest

ListTopicsDetectionJobsResponse

OutputDataConfig

Provides configuration parameters for the output of topic detection jobs.

PartOfSpeechTag
Identifies the part of speech represented by the token and gives the confidence that Amazon Comprehend has that the part of speech was correctly identified. For more information about the parts of speech that Amazon Comprehend can identify, see how-syntax.

SentimentDetectionJobFilter
Provides information for filtering a list of dominant language detection jobs. For more information, see the operation.

SentimentDetectionJobProperties
Provides information about a sentiment detection job.

SentimentScore
Describes the level of confidence that Amazon Comprehend has in the accuracy of its detection of sentiments.

StartDominantLanguageDetectionJobRequest

StartDominantLanguageDetectionJobResponse

StartEntitiesDetectionJobRequest

StartEntitiesDetectionJobResponse

StartKeyPhrasesDetectionJobRequest

StartKeyPhrasesDetectionJobResponse

StartSentimentDetectionJobRequest

StartSentimentDetectionJobResponse

StartTopicsDetectionJobRequest

StartTopicsDetectionJobResponse

StopDominantLanguageDetectionJobRequest

StopDominantLanguageDetectionJobResponse

StopEntitiesDetectionJobRequest

StopEntitiesDetectionJobResponse

StopKeyPhrasesDetectionJobRequest

StopKeyPhrasesDetectionJobResponse

StopSentimentDetectionJobRequest

StopSentimentDetectionJobResponse

SyntaxToken
Represents a work in the input text that was recognized and assigned a part of speech. There is one syntax token record for each word in the source text.

TopicsDetectionJobFilter
Provides information for filtering topic detection jobs. For more information, see .

TopicsDetectionJobProperties
Provides information about a topic detection job.

Enums

BatchDetectDominantLanguageError
Errors returned by BatchDetectDominantLanguage

BatchDetectEntitiesError
Errors returned by BatchDetectEntities

BatchDetectKeyPhrasesError
Errors returned by BatchDetectKeyPhrases

BatchDetectSentimentError
Errors returned by BatchDetectSentiment

BatchDetectSyntaxError
Errors returned by BatchDetectSyntax

DescribeDominantLanguageDetectionJobError
Errors returned by DescribeDominantLanguageDetectionJob

DescribeEntitiesDetectionJobError
Errors returned by DescribeEntitiesDetectionJob

DescribeKeyPhrasesDetectionJobError
Errors returned by DescribeKeyPhrasesDetectionJob

DescribeSentimentDetectionJobError
Errors returned by DescribeSentimentDetectionJob

DescribeTopicsDetectionJobError
Errors returned by DescribeTopicsDetectionJob

DetectDominantLanguageError
Errors returned by DetectDominantLanguage

DetectEntitiesError
Errors returned by DetectEntities

DetectKeyPhrasesError
Errors returned by DetectKeyPhrases

DetectSentimentError
Errors returned by DetectSentiment

DetectSyntaxError
Errors returned by DetectSyntax

ListDominantLanguageDetectionJobsError
Errors returned by ListDominantLanguageDetectionJobs

ListEntitiesDetectionJobsError
Errors returned by ListEntitiesDetectionJobs

ListKeyPhrasesDetectionJobsError
Errors returned by ListKeyPhrasesDetectionJobs

ListSentimentDetectionJobsError
Errors returned by ListSentimentDetectionJobs

ListTopicsDetectionJobsError
Errors returned by ListTopicsDetectionJobs

StartDominantLanguageDetectionJobError
Errors returned by StartDominantLanguageDetectionJob

StartEntitiesDetectionJobError
Errors returned by StartEntitiesDetectionJob

StartKeyPhrasesDetectionJobError
Errors returned by StartKeyPhrasesDetectionJob

StartSentimentDetectionJobError
Errors returned by StartSentimentDetectionJob

StartTopicsDetectionJobError
Errors returned by StartTopicsDetectionJob

StopDominantLanguageDetectionJobError
Errors returned by StopDominantLanguageDetectionJob

StopEntitiesDetectionJobError
Errors returned by StopEntitiesDetectionJob

StopKeyPhrasesDetectionJobError
Errors returned by StopKeyPhrasesDetectionJob

StopSentimentDetectionJobError
Errors returned by StopSentimentDetectionJob

Traits

Comprehend
Trait representing the capabilities of the Amazon Comprehend API. Amazon Comprehend clients implement this trait.