Struct rusoto_comprehend::ComprehendClient

pub struct ComprehendClient { /* fields omitted */ }

A client for the Amazon Comprehend API.

Methods

impl ComprehendClient

pub fn new(region: Region) -> ComprehendClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ComprehendClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,