Trait rusoto_comprehend::Comprehend

pub trait Comprehend {
    fn batch_detect_dominant_language(
        &self, 
        input: BatchDetectDominantLanguageRequest
    ) -> RusotoFuture<BatchDetectDominantLanguageResponse, BatchDetectDominantLanguageError>;

    fn batch_detect_entities(
        &self, 
        input: BatchDetectEntitiesRequest
    ) -> RusotoFuture<BatchDetectEntitiesResponse, BatchDetectEntitiesError>;

    fn batch_detect_key_phrases(
        &self, 
        input: BatchDetectKeyPhrasesRequest
    ) -> RusotoFuture<BatchDetectKeyPhrasesResponse, BatchDetectKeyPhrasesError>;

    fn batch_detect_sentiment(
        &self, 
        input: BatchDetectSentimentRequest
    ) -> RusotoFuture<BatchDetectSentimentResponse, BatchDetectSentimentError>;

    fn batch_detect_syntax(
        &self, 
        input: BatchDetectSyntaxRequest
    ) -> RusotoFuture<BatchDetectSyntaxResponse, BatchDetectSyntaxError>;

    fn describe_dominant_language_detection_job(
        &self, 
        input: DescribeDominantLanguageDetectionJobRequest
    ) -> RusotoFuture<DescribeDominantLanguageDetectionJobResponse, DescribeDominantLanguageDetectionJobError>;

    fn describe_entities_detection_job(
        &self, 
        input: DescribeEntitiesDetectionJobRequest
    ) -> RusotoFuture<DescribeEntitiesDetectionJobResponse, DescribeEntitiesDetectionJobError>;

    fn describe_key_phrases_detection_job(
        &self, 
        input: DescribeKeyPhrasesDetectionJobRequest
    ) -> RusotoFuture<DescribeKeyPhrasesDetectionJobResponse, DescribeKeyPhrasesDetectionJobError>;

    fn describe_sentiment_detection_job(
        &self, 
        input: DescribeSentimentDetectionJobRequest
    ) -> RusotoFuture<DescribeSentimentDetectionJobResponse, DescribeSentimentDetectionJobError>;

    fn describe_topics_detection_job(
        &self, 
        input: DescribeTopicsDetectionJobRequest
    ) -> RusotoFuture<DescribeTopicsDetectionJobResponse, DescribeTopicsDetectionJobError>;

    fn detect_dominant_language(
        &self, 
        input: DetectDominantLanguageRequest
    ) -> RusotoFuture<DetectDominantLanguageResponse, DetectDominantLanguageError>;

    fn detect_entities(
        &self, 
        input: DetectEntitiesRequest
    ) -> RusotoFuture<DetectEntitiesResponse, DetectEntitiesError>;

    fn detect_key_phrases(
        &self, 
        input: DetectKeyPhrasesRequest
    ) -> RusotoFuture<DetectKeyPhrasesResponse, DetectKeyPhrasesError>;

    fn detect_sentiment(
        &self, 
        input: DetectSentimentRequest
    ) -> RusotoFuture<DetectSentimentResponse, DetectSentimentError>;

    fn detect_syntax(
        &self, 
        input: DetectSyntaxRequest
    ) -> RusotoFuture<DetectSyntaxResponse, DetectSyntaxError>;

    fn list_dominant_language_detection_jobs(
        &self, 
        input: ListDominantLanguageDetectionJobsRequest
    ) -> RusotoFuture<ListDominantLanguageDetectionJobsResponse, ListDominantLanguageDetectionJobsError>;

    fn list_entities_detection_jobs(
        &self, 
        input: ListEntitiesDetectionJobsRequest
    ) -> RusotoFuture<ListEntitiesDetectionJobsResponse, ListEntitiesDetectionJobsError>;

    fn list_key_phrases_detection_jobs(
        &self, 
        input: ListKeyPhrasesDetectionJobsRequest
    ) -> RusotoFuture<ListKeyPhrasesDetectionJobsResponse, ListKeyPhrasesDetectionJobsError>;

    fn list_sentiment_detection_jobs(
        &self, 
        input: ListSentimentDetectionJobsRequest
    ) -> RusotoFuture<ListSentimentDetectionJobsResponse, ListSentimentDetectionJobsError>;

    fn list_topics_detection_jobs(
        &self, 
        input: ListTopicsDetectionJobsRequest
    ) -> RusotoFuture<ListTopicsDetectionJobsResponse, ListTopicsDetectionJobsError>;

    fn start_dominant_language_detection_job(
        &self, 
        input: StartDominantLanguageDetectionJobRequest
    ) -> RusotoFuture<StartDominantLanguageDetectionJobResponse, StartDominantLanguageDetectionJobError>;

    fn start_entities_detection_job(
        &self, 
        input: StartEntitiesDetectionJobRequest
    ) -> RusotoFuture<StartEntitiesDetectionJobResponse, StartEntitiesDetectionJobError>;

    fn start_key_phrases_detection_job(
        &self, 
        input: StartKeyPhrasesDetectionJobRequest
    ) -> RusotoFuture<StartKeyPhrasesDetectionJobResponse, StartKeyPhrasesDetectionJobError>;

    fn start_sentiment_detection_job(
        &self, 
        input: StartSentimentDetectionJobRequest
    ) -> RusotoFuture<StartSentimentDetectionJobResponse, StartSentimentDetectionJobError>;

    fn start_topics_detection_job(
        &self, 
        input: StartTopicsDetectionJobRequest
    ) -> RusotoFuture<StartTopicsDetectionJobResponse, StartTopicsDetectionJobError>;

    fn stop_dominant_language_detection_job(
        &self, 
        input: StopDominantLanguageDetectionJobRequest
    ) -> RusotoFuture<StopDominantLanguageDetectionJobResponse, StopDominantLanguageDetectionJobError>;

    fn stop_entities_detection_job(
        &self, 
        input: StopEntitiesDetectionJobRequest
    ) -> RusotoFuture<StopEntitiesDetectionJobResponse, StopEntitiesDetectionJobError>;

    fn stop_key_phrases_detection_job(
        &self, 
        input: StopKeyPhrasesDetectionJobRequest
    ) -> RusotoFuture<StopKeyPhrasesDetectionJobResponse, StopKeyPhrasesDetectionJobError>;

    fn stop_sentiment_detection_job(
        &self, 
        input: StopSentimentDetectionJobRequest
    ) -> RusotoFuture<StopSentimentDetectionJobResponse, StopSentimentDetectionJobError>;
}

Trait representing the capabilities of the Amazon Comprehend API. Amazon Comprehend clients implement this trait.

Required Methods

fn batch_detect_dominant_language(
&self,
input: BatchDetectDominantLanguageRequest
) -> RusotoFuture<BatchDetectDominantLanguageResponse, BatchDetectDominantLanguageError>

Determines the dominant language of the input text for a batch of documents. For a list of languages that Amazon Comprehend can detect, see Amazon Comprehend Supported Languages.

fn batch_detect_entities(
&self,
input: BatchDetectEntitiesRequest
) -> RusotoFuture<BatchDetectEntitiesResponse, BatchDetectEntitiesError>

Inspects the text of a batch of documents for named entities and returns information about them. For more information about named entities, see how-entities

fn batch_detect_key_phrases(
&self,
input: BatchDetectKeyPhrasesRequest
) -> RusotoFuture<BatchDetectKeyPhrasesResponse, BatchDetectKeyPhrasesError>

Detects the key noun phrases found in a batch of documents.

fn batch_detect_sentiment(
&self,
input: BatchDetectSentimentRequest
) -> RusotoFuture<BatchDetectSentimentResponse, BatchDetectSentimentError>

Inspects a batch of documents and returns an inference of the prevailing sentiment, POSITIVE, NEUTRAL, MIXED, or NEGATIVE, in each one.

fn batch_detect_syntax(
&self,
input: BatchDetectSyntaxRequest
) -> RusotoFuture<BatchDetectSyntaxResponse, BatchDetectSyntaxError>

Inspects the text of a batch of documents for the syntax and part of speech of the words in the document and returns information about them. For more information, see how-syntax.

fn describe_dominant_language_detection_job(
&self,
input: DescribeDominantLanguageDetectionJobRequest
) -> RusotoFuture<DescribeDominantLanguageDetectionJobResponse, DescribeDominantLanguageDetectionJobError>

Gets the properties associated with a dominant language detection job. Use this operation to get the status of a detection job.

fn describe_entities_detection_job(
&self,
input: DescribeEntitiesDetectionJobRequest
) -> RusotoFuture<DescribeEntitiesDetectionJobResponse, DescribeEntitiesDetectionJobError>

Gets the properties associated with an entities detection job. Use this operation to get the status of a detection job.

fn describe_key_phrases_detection_job(
&self,
input: DescribeKeyPhrasesDetectionJobRequest
) -> RusotoFuture<DescribeKeyPhrasesDetectionJobResponse, DescribeKeyPhrasesDetectionJobError>

Gets the properties associated with a key phrases detection job. Use this operation to get the status of a detection job.

fn describe_sentiment_detection_job(
&self,
input: DescribeSentimentDetectionJobRequest
) -> RusotoFuture<DescribeSentimentDetectionJobResponse, DescribeSentimentDetectionJobError>

Gets the properties associated with a sentiment detection job. Use this operation to get the status of a detection job.

fn describe_topics_detection_job(
&self,
input: DescribeTopicsDetectionJobRequest
) -> RusotoFuture<DescribeTopicsDetectionJobResponse, DescribeTopicsDetectionJobError>

Gets the properties associated with a topic detection job. Use this operation to get the status of a detection job.

fn detect_dominant_language(
&self,
input: DetectDominantLanguageRequest
) -> RusotoFuture<DetectDominantLanguageResponse, DetectDominantLanguageError>

Determines the dominant language of the input text. For a list of languages that Amazon Comprehend can detect, see Amazon Comprehend Supported Languages.

fn detect_entities(
&self,
input: DetectEntitiesRequest
) -> RusotoFuture<DetectEntitiesResponse, DetectEntitiesError>

Inspects text for named entities, and returns information about them. For more information, about named entities, see how-entities.

fn detect_key_phrases(
&self,
input: DetectKeyPhrasesRequest
) -> RusotoFuture<DetectKeyPhrasesResponse, DetectKeyPhrasesError>

Detects the key noun phrases found in the text.

fn detect_sentiment(
&self,
input: DetectSentimentRequest
) -> RusotoFuture<DetectSentimentResponse, DetectSentimentError>

Inspects text and returns an inference of the prevailing sentiment (POSITIVE, NEUTRAL, MIXED, or NEGATIVE).

fn detect_syntax(
&self,
input: DetectSyntaxRequest
) -> RusotoFuture<DetectSyntaxResponse, DetectSyntaxError>

Inspects text for syntax and the part of speech of words in the document. For more information, how-syntax.

fn list_dominant_language_detection_jobs(
&self,
input: ListDominantLanguageDetectionJobsRequest
) -> RusotoFuture<ListDominantLanguageDetectionJobsResponse, ListDominantLanguageDetectionJobsError>

Gets a list of the dominant language detection jobs that you have submitted.

fn list_entities_detection_jobs(
&self,
input: ListEntitiesDetectionJobsRequest
) -> RusotoFuture<ListEntitiesDetectionJobsResponse, ListEntitiesDetectionJobsError>

Gets a list of the entity detection jobs that you have submitted.

fn list_key_phrases_detection_jobs(
&self,
input: ListKeyPhrasesDetectionJobsRequest
) -> RusotoFuture<ListKeyPhrasesDetectionJobsResponse, ListKeyPhrasesDetectionJobsError>

Get a list of key phrase detection jobs that you have submitted.

fn list_sentiment_detection_jobs(
&self,
input: ListSentimentDetectionJobsRequest
) -> RusotoFuture<ListSentimentDetectionJobsResponse, ListSentimentDetectionJobsError>

Gets a list of sentiment detection jobs that you have submitted.

fn list_topics_detection_jobs(
&self,
input: ListTopicsDetectionJobsRequest
) -> RusotoFuture<ListTopicsDetectionJobsResponse, ListTopicsDetectionJobsError>

Gets a list of the topic detection jobs that you have submitted.

fn start_dominant_language_detection_job(
&self,
input: StartDominantLanguageDetectionJobRequest
) -> RusotoFuture<StartDominantLanguageDetectionJobResponse, StartDominantLanguageDetectionJobError>

Starts an asynchronous dominant language detection job for a collection of documents. Use the operation to track the status of a job.

fn start_entities_detection_job(
&self,
input: StartEntitiesDetectionJobRequest
) -> RusotoFuture<StartEntitiesDetectionJobResponse, StartEntitiesDetectionJobError>

Starts an asynchronous entity detection job for a collection of documents. Use the operation to track the status of a job.

fn start_key_phrases_detection_job(
&self,
input: StartKeyPhrasesDetectionJobRequest
) -> RusotoFuture<StartKeyPhrasesDetectionJobResponse, StartKeyPhrasesDetectionJobError>

Starts an asynchronous key phrase detection job for a collection of documents. Use the operation to track the status of a job.

fn start_sentiment_detection_job(
&self,
input: StartSentimentDetectionJobRequest
) -> RusotoFuture<StartSentimentDetectionJobResponse, StartSentimentDetectionJobError>

Starts an asynchronous sentiment detection job for a collection of documents. use the operation to track the status of a job.

fn start_topics_detection_job(
&self,
input: StartTopicsDetectionJobRequest
) -> RusotoFuture<StartTopicsDetectionJobResponse, StartTopicsDetectionJobError>

Starts an asynchronous topic detection job. Use the DescribeTopicDetectionJob operation to track the status of a job.

fn stop_dominant_language_detection_job(
&self,
input: StopDominantLanguageDetectionJobRequest
) -> RusotoFuture<StopDominantLanguageDetectionJobResponse, StopDominantLanguageDetectionJobError>

Stops a dominant language detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

fn stop_entities_detection_job(
&self,
input: StopEntitiesDetectionJobRequest
) -> RusotoFuture<StopEntitiesDetectionJobResponse, StopEntitiesDetectionJobError>

Stops an entities detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

fn stop_key_phrases_detection_job(
&self,
input: StopKeyPhrasesDetectionJobRequest
) -> RusotoFuture<StopKeyPhrasesDetectionJobResponse, StopKeyPhrasesDetectionJobError>

Stops a key phrases detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.

fn stop_sentiment_detection_job(
&self,
input: StopSentimentDetectionJobRequest
) -> RusotoFuture<StopSentimentDetectionJobResponse, StopSentimentDetectionJobError>

Stops a sentiment detection job in progress.

If the job state is IN_PROGRESS the job is marked for termination and put into the STOP_REQUESTED state. If the job completes before it can be stopped, it is put into the COMPLETED state; otherwise the job is be stopped and put into the STOPPED state.

If the job is in the COMPLETED or FAILED state when you call the StopDominantLanguageDetectionJob operation, the operation returns a 400 Internal Request Exception.

When a job is stopped, any documents already processed are written to the output location.